👨 Add profile functionality

+++ controllers/sign-up.js
@@ -4,7 +4,13 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 exports.getSignUpPage = (req, res, next) => {
-  res.render('sign-up', {errors: []});
+  res.render('sign-up', {      
+    firstName: '',
+    lastName: '',
+    email: '',
+    phoneNumber: '',
+    errors: []
+  });
 };
 
 exports.postSignUpPage = async (req, res, next) => {
@@ -12,8 +18,14 @@ exports.postSignUpPage = async (req, res, next) => {
   const { firstName, lastName, email, phoneNumber, password} = req.body;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    await fs.unlink(path.join('.', req.file.path));
+    if (req.file) {
+      await fs.unlink(path.join('.', req.file.path));
+    }
     return res.render('sign-up', {
+      firstName,
+      lastName,
+      email,
+      phoneNumber,
       errors: errors.array(),
     });
   }
@@ -22,7 +34,7 @@ exports.postSignUpPage = async (req, res, next) => {
   if (req.file) {
     imageURL = '/' + req.file.path;
   }
-  const user = new User({firstName, lastName, email, phoneNum: phoneNumber, password, imageURL});
+  const user = new User({firstName, lastName, email, phoneNum: phoneNumber, password, img_url: imageURL});
   await user.save(); // Saves in database
   req.session.user = user;
   res.redirect('/');
